+++ PyCrypCli/commands/miner.py
@@ -1,10 +1,10 @@
 from typing import List
 
-from commands.command import command, CTX_DEVICE
-from exceptions import *
-from game import Game
-from game_objects import Service, Miner
-from util import is_uuid
+from ..commands.command import command, CTX_DEVICE
+from ..exceptions import *
+from ..game import Game
+from ..game_objects import Service, Miner
+from ..util import is_uuid
 
 
 @command(["miner"], CTX_DEVICE, "Manager your Morphcoin miners")
