+++ PyCrypCli/exceptions.py
