+++ PyCrypCli/game_objects.py
