+++ PyCrypCli/util.py
